requirements.txt
streamlit>=1.28.0
edge-tts>=6.1.9
langchain-openai>=0.0.2
tavily-python>=0.3.0
python-dotenv>=1.0.0
